refactor(nav): modernize backward loop, split parent-filter tests
- update_navigation.go: use slices.Backward in moveCursor's
  expanded-group lookup to satisfy the modernize linter introduced
  in golangci-lint v2.12.0.
- Move TestNavigateParent_ClearsFilter and
  TestNavigateParentFromResourceTypes_FilterClearRevealsContexts into
  update_navigation_parent_filter_test.go to keep test files within
  the project's 800-line guideline (CodeRabbit review feedback).
